+++ org.ektorp/src/main/java/org/ektorp/CouchDbInstance.java
@@ -23,6 +23,7 @@ public interface CouchDbInstance {
 	 * @return true if the database exists.
 	 */
 	boolean checkIfDbExists(DbPath db);
+	boolean checkIfDbExists(String path);
 
 	void createDatabase(DbPath path);
 	void createDatabase(String path);
@@ -101,4 +102,8 @@ public interface CouchDbInstance {
      */
     String deleteConfiguration(String section, String key);
 
+    /**
+     * @return all active tasks
+     */
+    Collection<ActiveTask> getActiveTasks();
 }