+++ org.ektorp/src/main/java/org/ektorp/changes/DocumentChange.java
@@ -1,6 +1,9 @@
 package org.ektorp.changes;
 
-import org.codehaus.jackson.*;
+import com.fasterxml.jackson.databind.JsonNode;
+
+import java.util.List;
+
 /**
  * Represents a document change within a database.
  * @author henrik lundgren
@@ -12,17 +15,25 @@ public interface DocumentChange {
 	 */
 	int getSequence();
 	/**
-	 * 
+	 *
 	 * @return the id of the changed document.
 	 */
 	String getId();
 	/**
-	 * 
+	 *
 	 * @return the revision the document had at the time of change.
 	 */
 	String getRevision();
+
+    /**
+     *
+     * @return the collection of revisions, including conflicts
+     * @since 1.4.0
+     */
+    public List<String> getRevisions();
+
 	/**
-	 * 
+	 *
 	 * @return true if the changed document has been deleted.
 	 */
 	boolean isDeleted();