+++ org.ektorp/src/main/java/org/ektorp/impl/StdCouchDbInstance.java
@@ -5,14 +5,15 @@ import static java.lang.String.*;
 import java.io.*;
 import java.util.*;
 
-import org.codehaus.jackson.map.*;
-import org.codehaus.jackson.type.*;
+import com.fasterxml.jackson.core.type.TypeReference;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
 import org.ektorp.*;
 import org.ektorp.http.*;
 import org.ektorp.util.*;
 import org.slf4j.*;
 /**
- * 
+ *
  * @author henrik lundgren
  *
  */
@@ -20,7 +21,7 @@ public class StdCouchDbInstance implements CouchDbInstance {
 
 	private final static Logger LOG = LoggerFactory.getLogger(StdCouchDbInstance.class);
 	private final static TypeReference<List<String>> STRING_LIST_TYPE_DEF = new TypeReference<List<String>>() {};
-	
+
 	private final HttpClient client;
 	private final RestTemplate restTemplate;
 	private final ObjectMapper objectMapper;
@@ -30,7 +31,7 @@ public class StdCouchDbInstance implements CouchDbInstance {
 	public StdCouchDbInstance(HttpClient client) {
 		this(client, new StdObjectMapperFactory());
 	}
-	
+
 	public StdCouchDbInstance(HttpClient client, ObjectMapperFactory of) {
 		Assert.notNull(client, "HttpClient may not be null");
 		Assert.notNull(of, "ObjectMapperFactory may not be null");
@@ -40,11 +41,11 @@ public class StdCouchDbInstance implements CouchDbInstance {
 		this.jsonSerializer = new StreamingJsonSerializer(objectMapper);
 		this.objectMapperFactory = of;
 	}
-	
+
 	public void createDatabase(String path) {
 		createDatabase(DbPath.fromString(path));
 	}
-	
+
 	public void createDatabase(DbPath db) {
 		if (checkIfDbExists(db)) {
 			throw new DbAccessException(format("A database with path %s already exists", db.getPath()));
@@ -56,6 +57,11 @@ public class StdCouchDbInstance implements CouchDbInstance {
 	public void deleteDatabase(String path) {
 		Assert.notNull(path);
 		restTemplate.delete(DbPath.fromString(path).getPath());
+	}
+	
+	@Override
+	public boolean checkIfDbExists(String path) {
+	    return checkIfDbExists(DbPath.fromString(path));
 	}
 
 	@Override
@@ -80,7 +86,7 @@ public class StdCouchDbInstance implements CouchDbInstance {
 			}
 		});
 	}
-	
+
 	public ReplicationStatus replicate(ReplicationCommand cmd) {
 		try {
 			return restTemplate.post("/_replicate", objectMapper.writeValueAsString(cmd), new StdResponseHandler<ReplicationStatus>() {
@@ -89,7 +95,7 @@ public class StdCouchDbInstance implements CouchDbInstance {
 						throws Exception {
 					return objectMapper.readValue(hr.getContent(), ReplicationStatus.class);
 				}
-			});	
+			});
 		} catch (IOException e) {
 			throw Exceptions.propagate(e);
 		}
@@ -98,7 +104,7 @@ public class StdCouchDbInstance implements CouchDbInstance {
 	public HttpClient getConnection() {
 		return client;
 	}
-	
+
 	public CouchDbConnector createConnector(String path,
 			boolean createIfNotExists) {
 		CouchDbConnector db = new StdCouchDbConnector(path, this, objectMapperFactory);
@@ -155,8 +161,7 @@ public class StdCouchDbInstance implements CouchDbInstance {
          new StdResponseHandler<String>() {
             @Override
             public String success(HttpResponse hr) throws Exception {
-               String s = objectMapper.readValue(hr.getContent(), String.class);
-               return s;
+               return objectMapper.readValue(hr.getContent(), String.class);
             }
          });
     }
@@ -170,9 +175,29 @@ public class StdCouchDbInstance implements CouchDbInstance {
           new StdResponseHandler<String>() {
              @Override
              public String success(HttpResponse hr) throws Exception {
-                String s = objectMapper.readValue(hr.getContent(), String.class);
-                return s;
+                return objectMapper.readValue(hr.getContent(), String.class);
              }
          });
+   }
+
+   @Override
+   public Collection<ActiveTask> getActiveTasks() {
+      String url = "/_active_tasks";
+      List<StdActiveTask> tasks = restTemplate.get(url,
+         new StdResponseHandler<List<StdActiveTask>>() {
+         @Override
+         public List<StdActiveTask> success(HttpResponse hr) throws Exception {
+            return objectMapper.readValue(hr.getContent(), new TypeReference<List<StdActiveTask>>() {});
+         }
+      });
+
+      // We have to copy the list here because Java lacks covariance (i.e. we can't just return
+      // the List<StdActiveTask> because it's not a Collection<ActiveTask>).
+      Collection<ActiveTask> ret = new ArrayList<ActiveTask>();
+      for (StdActiveTask task : tasks) {
+          ret.add(task);
+      }
+
+      return ret;
    }
 }