+++ org.ektorp/src/main/java/org/ektorp/impl/changes/StdDocumentChange.java
@@ -1,10 +1,15 @@
 package org.ektorp.impl.changes;
 
-import org.codehaus.jackson.*;
+import com.fasterxml.jackson.databind.JsonNode;
 import org.ektorp.changes.*;
 import org.ektorp.util.*;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
 /**
- * 
+ *
  * @author henrik lundgren
  *
  */
@@ -15,38 +20,39 @@ public class StdDocumentChange implements DocumentChange {
 	private static final String ID_FIELD_NAME = "id";
 	private static final String DOC_FIELD_NAME = "doc";
 	private static final String DELETED_FIELD_NAME = "deleted";
-	
-	private final JsonNode node;
-	
-	public StdDocumentChange(JsonNode node) {
+    private static final String CHANGES_FIELD_NAME = "changes";
+
+    private final JsonNode node;
+
+    public StdDocumentChange(JsonNode node) {
 		Assert.notNull(node, "node may not be null");
 		this.node = node;
 	}
-	
+
 	public int getSequence() {
-		return node.get(SEQ_FIELD_NAME).getIntValue();
+		return node.get(SEQ_FIELD_NAME).intValue();
 	}
-	
+
 	public String getId() {
-		return node.get(ID_FIELD_NAME).getTextValue();
+		return node.get(ID_FIELD_NAME).textValue();
 	}
-	
+
 	public boolean isDeleted() {
-		JsonNode deleted = node.findPath(DELETED_FIELD_NAME); 
-		return !deleted.isMissingNode() && deleted.getBooleanValue();
+		JsonNode deleted = node.findPath(DELETED_FIELD_NAME);
+		return !deleted.isMissingNode() && deleted.booleanValue();
 	}
-	
+
 	public String getDoc() {
 		return nodeAsString(node.findValue(DOC_FIELD_NAME));
 	}
-	
+
 	public JsonNode getDocAsNode() {
 		return node.findPath(DOC_FIELD_NAME);
 	}
-	
+
 	private String nodeAsString(JsonNode node) {
 		if (isNull(node)) return null;
-		return node.isContainerNode() ? node.toString() : node.getValueAsText();
+		return node.isContainerNode() ? node.toString() : node.asText();
 	}
 
 	private boolean isNull(JsonNode node) {
@@ -54,9 +60,17 @@ public class StdDocumentChange implements DocumentChange {
 	}
 
 	public String getRevision() {
-		return nodeAsString(node.findPath(REV_FIELD_NAME));
+        return nodeAsString(node.findValue(REV_FIELD_NAME));
 	}
-	
+
+    public List<String> getRevisions() {
+        List<String> revisions = new ArrayList<String>();
+        for (JsonNode changesNode : node.get(CHANGES_FIELD_NAME)) {
+            revisions.add(nodeAsString(changesNode.get(REV_FIELD_NAME)));
+        }
+        return Collections.unmodifiableList(revisions);
+    }
+
 	@Override
 	public String toString() {
 		return node.toString();