+++ org.ektorp/src/main/java/org/ektorp/impl/docref/BackReferencedBeanDeserializer.java
@@ -4,9 +4,14 @@ import java.io.*;
 import java.lang.reflect.*;
 import java.util.*;
 
-import org.codehaus.jackson.*;
-import org.codehaus.jackson.map.*;
-import org.codehaus.jackson.map.deser.*;
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.DeserializationConfig;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JsonMappingException;
+import com.fasterxml.jackson.databind.deser.BeanDeserializer;
+import com.fasterxml.jackson.databind.deser.ResolvableDeserializer;
+import com.fasterxml.jackson.databind.deser.std.StdDeserializer;
 import org.ektorp.*;
 import org.ektorp.docref.*;
 import org.ektorp.util.*;
@@ -23,18 +28,15 @@ public class BackReferencedBeanDeserializer extends StdDeserializer<Object>
 	private final BeanDeserializer delegate;
 	private final List<ConstructibleAnnotatedCollection> backReferencedFields;
 	private final Class<?> clazz;
-	private final ObjectMapper objectMapper;
 
 	public BackReferencedBeanDeserializer(BeanDeserializer deserializer,
 			List<ConstructibleAnnotatedCollection> fields,
-			CouchDbConnector couchDbConnector, Class<?> clazz,
-			ObjectMapper objectMapper) {
+			CouchDbConnector couchDbConnector, Class<?> clazz) {
 		super(clazz);
 		this.clazz = clazz;
 		this.delegate = deserializer;
 		this.couchDbConnector = couchDbConnector;
 		this.backReferencedFields = fields;
-		this.objectMapper = objectMapper;
 	}
 
 	@Override
@@ -58,12 +60,11 @@ public class BackReferencedBeanDeserializer extends StdDeserializer<Object>
 
 				if (ann.fetch().equals(FetchType.EAGER)) {
 					handler = new ViewBasedCollection(id, couchDbConnector,
-							clazz, ann, constructibleField, objectMapper);
+							clazz, ann, constructibleField);
 					handler.initialize();
 				} else {
 					handler = new LazyLoadingViewBasedCollection(id,
-							couchDbConnector, clazz, ann, constructibleField,
-							objectMapper);
+							couchDbConnector, clazz, ann, constructibleField);
 
 				}
 
@@ -89,9 +90,9 @@ public class BackReferencedBeanDeserializer extends StdDeserializer<Object>
 		return deserializedObject;
 	}
 
-	public void resolve(DeserializationConfig config,
-			DeserializerProvider provider) throws JsonMappingException {
-		delegate.resolve(config, provider);
+	@Override
+	public void resolve(DeserializationContext ctxt) throws JsonMappingException {
+		delegate.resolve(ctxt);
 	}
 
 }