+++ org.ektorp/src/main/java/org/ektorp/impl/jackson/EktorpAnnotationIntrospector.java
@@ -9,50 +9,56 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
-import org.codehaus.jackson.map.introspect.AnnotatedClass;
-import org.codehaus.jackson.map.introspect.AnnotatedField;
-import org.codehaus.jackson.map.introspect.AnnotatedMember;
-import org.codehaus.jackson.map.introspect.AnnotatedMethod;
-import org.codehaus.jackson.map.introspect.NopAnnotationIntrospector;
 import org.ektorp.docref.DocumentReferences;
 import org.ektorp.impl.NameConventions;
 import org.ektorp.util.Predicate;
 import org.ektorp.util.ReflectionUtils;
+
+import com.fasterxml.jackson.databind.introspect.Annotated;
+import com.fasterxml.jackson.databind.introspect.AnnotatedClass;
+import com.fasterxml.jackson.databind.introspect.AnnotatedField;
+import com.fasterxml.jackson.databind.introspect.AnnotatedMember;
+import com.fasterxml.jackson.databind.introspect.AnnotatedMethod;
+import com.fasterxml.jackson.databind.introspect.NopAnnotationIntrospector;
 
 public class EktorpAnnotationIntrospector extends NopAnnotationIntrospector {
 
 	private final Map<Class<?>, Set<String>> ignorableMethods = new HashMap<Class<?>, Set<String>>();
 	private final Set<Class<?>> annotatedClasses = new HashSet<Class<?>>();
 
-	
+
 	@Override
 	public boolean isHandled(Annotation ann) {
 		return DocumentReferences.class == ann.annotationType();
 	}
-	
+
 	@Override
 	public boolean hasIgnoreMarker(AnnotatedMember member) {
-		boolean b = super.hasIgnoreMarker(member); 
-		return b;
+		return super.hasIgnoreMarker(member);
 	}
-	
-	@Override
+
 	public boolean isIgnorableField(AnnotatedField f) {
 		return f.hasAnnotation(DocumentReferences.class);
 	}
-	
-	@Override
+
 	public boolean isIgnorableMethod(AnnotatedMethod m) {
 		Set<String> names = ignorableMethods.get(m.getDeclaringClass());
 		if (names == null) {
 			initIgnorableMethods(m.getDeclaringClass());
 			names = ignorableMethods.get(m.getDeclaringClass());
 		}
-		
+
 		return names.contains(m.getName());
 	}
 
-    @Override
+	@Override
+	public String[] findPropertiesToIgnore(Annotated ac) {
+		if(ac instanceof AnnotatedClass){
+			return findPropertiesToIgnore((AnnotatedClass) ac);
+		}
+		return super.findPropertiesToIgnore(ac);
+	}
+	
     public String[] findPropertiesToIgnore(AnnotatedClass ac) {
     	List<String> ignoreFields = null;
     	for (AnnotatedField f : ac.fields()) {
@@ -65,7 +71,7 @@ public class EktorpAnnotationIntrospector extends NopAnnotationIntrospector {
     	}
         return ignoreFields != null ? ignoreFields.toArray(new String[ignoreFields.size()]) : null;
     }
-	
+
 	private void initIgnorableMethods(final Class<?> clazz) {
 		final Set<String> names = new HashSet<String>();
 		ReflectionUtils.eachField(clazz, new Predicate<Field>() {
@@ -80,5 +86,5 @@ public class EktorpAnnotationIntrospector extends NopAnnotationIntrospector {
 		});
 		ignorableMethods.put(clazz, names);
 	}
-	
+
 }